Add delete route and reworded code for better readability
