+++ backend/app/schemas/routers.py
@@ -1,10 +1,5 @@
-from typing import List, Literal, Optional
+from typing import List, Literal
 from pydantic import BaseModel, Field, constr
-
-
-class SequenceInput(BaseModel):
-    id: constr(strip_whitespace=True, min_length=1)
-    sequence: constr(strip_whitespace=True, min_length=1)
 
 
 class ModelConfig(BaseModel):
@@ -13,41 +8,6 @@ class ModelConfig(BaseModel):
     batch_size: int = Field(64, ge=1, le=1024)
     enable_ood: bool = False
     ood_threshold: float = Field(0.99, ge=0.0, le=1.0)
-
-
-class ClassificationRequest(BaseModel):
-    sequences: List[SequenceInput]
-    config: Optional[ModelConfig] = None
-    source: Optional[str] = None
-
-
-class SequenceResult(BaseModel):
-    id: int | None = None
-    sequence_id: str
-    length: int
-    gc_content: float
-    prediction: Literal["Virus", "Host", "Novel", "Uncertain", "Invalid"]
-    confidence: float
-    sequence_preview: str
-    organism_name: Optional[str] = None
-    explanation: Optional[str] = None
-    mahalanobis_distance: Optional[float] = None
-    energy_score: Optional[float] = None
-    ood_score: Optional[float] = None
-    uncertain: Optional[bool] = False
-    threshold_used: Optional[float] = None
-
-
-class ClassificationResponse(BaseModel):
-    total_sequences: int
-    virus_count: int
-    host_count: int
-    novel_count: int
-    uncertain_count: int
-    detailed_results: List[SequenceResult]
-    source: str
-    timestamp: str
-    processing_time: float
 
 
 class ChatMessage(BaseModel):
